add file padding
